+++ lib/toggle_reserves.m
@@ -80,9 +80,9 @@ end
 
 
 %%-----  ext2int  ------------------------------------------------------
-function mpc = userfcn_reserves_ext2int(mpc, args)
+function mpc = userfcn_reserves_ext2int(mpc, mpopt, args)
 %
-%   mpc = userfcn_reserves_ext2int(mpc, args)
+%   mpc = userfcn_reserves_ext2int(mpc, mpopt, args)
 %
 %   This is the 'ext2int' stage userfcn callback that prepares the input
 %   data for the formulation stage. It expects to find a 'reserves' field
@@ -197,9 +197,9 @@ om.add_quad_cost('Rcost', [], Cw, 0, {'R'});
 
 
 %%-----  int2ext  ------------------------------------------------------
-function results = userfcn_reserves_int2ext(results, args)
+function results = userfcn_reserves_int2ext(results, mpopt, args)
 %
-%   results = userfcn_reserves_int2ext(results, args)
+%   results = userfcn_reserves_int2ext(results, mpopt, args)
 %
 %   This is the 'int2ext' stage userfcn callback that converts everything
 %   back to external indexing and packages up the results. It expects to